+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 -   Various CSS fixes
 -   Fix client form
 -   Add missing all type to sort type filter
+-   Fix to update ending hour for appointment event
 
 # [0.5.3] (2025-17-01)